consumer-service/requirements.txt
pika==1.3.1
python-dotenv==1.0.0
SQLAlchemy==2.0.8
typing_extensions==4.5.0
